requirements.txt
astroid==2.0.2
colorama==0.3.9
dj-static==0.0.6
Django==2.0.7
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
paho-mqtt==1.3.1
pylint==2.1.0
pytz==2018.5
six==1.11.0
static3==0.7.0
typed-ast==1.1.0
typing==3.6.4
wrapt==1.10.11
